06_Multimodality/30_Image_Generation/production_service/requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
accelerate
Pillow
python-dotenv
pydantic
safetensors
